Fixed the commit verification failure using github copilot suggestions

+++ src/FundMe.sol
@@ -10,7 +10,7 @@ contract FundMe {
     uint256 public amount;
 
     address[] public funders;
-    mapping(address => uint) public addressToAmountFunded;
+    mapping(address => uint256) public addressToAmountFunded;
 
     constructor(address _priceFeed, uint256 _amount) {
         owner = payable(msg.sender);
@@ -24,15 +24,12 @@ contract FundMe {
     }
 
     function fund() public payable {
-        require(
-            msg.value.getConversionRate(priceFeed) >= amount,
-            "You need to spend more ETH!"
-        );
+        require(msg.value.getConversionRate(priceFeed) >= amount, "You need to spend more ETH!");
         addressToAmountFunded[msg.sender] += msg.value;
     }
 
     function withdraw() public onlyOwner {
-        (bool success, ) = owner.call{value: address(this).balance}("");
+        (bool success,) = owner.call{value: address(this).balance}("");
         require(success, "Transfer failed.");
         for (uint256 i = 0; i < funders.length; i++) {
             addressToAmountFunded[funders[i]] = 0;
